Update schema after bundle install

+++ db/schema.rb
@@ -38,7 +38,7 @@ ActiveRecord::Schema.define(version: 2022_04_06_201650) do
   end
 
   create_table "active_storage_variant_records", force: :cascade do |t|
-    t.bigint "blob_id", null: false
+    t.integer "blob_id", null: false
     t.string "variation_digest", null: false
     t.index ["blob_id", "variation_digest"], name: "index_active_storage_variant_records_uniqueness", unique: true
   end
